refactor: 자동차 이동거리 자료형 int 변경

+++ src/main/java/racingcar/car/Car.java
@@ -7,7 +7,7 @@ public abstract class Car {
     public Car() {
     }
 
-    public long getMoveDistance() {
+    public int getMoveDistance() {
         return moveDistance;
     }
